Create currency_formatter
In this we learnt about text package in java where we even created our own locale for india. We output the double precision payment in different currencies. We learnt about NumberFormat class in text package as well as locale class in util package. And also format method is used of NumberFormat class.