Create prime_check1
This was for checking prime number but the problem given was that there can be only one method named checkPrime but it can have any number of parameters and method overloading using more than 2 methods was also not allowed. So we used int... num(variable argument array) as a parameter and made another method to check prime or not as boolean.
One more error was their because of static use 'in' object inside InputStreamReader which was because "in" is static member of System class. Either we can write System.in there or we can import static java.lang.System.in   .