Create SingletonPattern1
Use of Singleton which restricts the instantiation of class to only one time and provides a global point of access to it.
Can be done in two ways:
Early Instantiation : creation of instance at load time which we did in above code
Lazy Instantiation : creation of instance when required.

Once submitted, our hidden Solution class will check your code by taking a String as input and then using your Singleton class to print a line.
So, it is not full code that can be compiled and run seperately. It can only run on hackerrank.